added FeedFetcher, PageDownloadTask, feed pins
The two new classes are used to download info from the various feeds.
Only the coupon one is implemented right now, but we can make it more
modular soon. Pins will appear on the map at the location of businesses
with coupons. We have not actually parsed out the individual coupons
yet... put those in a new window when you click on the pin? Snippet
should have number of coupons available?